Keep every touched file under 250 lines; dedup updater tests
Split the source and test modules that exceeded the 250-line limit into
smaller modules, preserving all public import paths and test behaviour:

- token_schemas.py -> facade re-exporting _enclosures/_tokens/_signs/_words
- fragments.py -> extract signs-resource factories into a sibling module
- test_fragment_updater, test_word, test_update_ocred_signs,
  test_cropped_annotations_service and test_transliterations_route split
  by group, with shared fixtures/helpers extracted into helper modules

Also remove the qlty:similar-code duplication in the fragment updater
tests (shared changelog helper plus parametrised field tests) and
gitignore local qlty/Claude tooling files.
