+++ .gitignore
@@ -140,3 +140,6 @@ core.*
 
 # Claude Code
 .claude/
+
+# qlty
+.qlty/
