+++ TASK-740-log.md
@@ -1,0 +1,59 @@
+# TASK-740 LOG — Fix qlty blocking issues
+
+## Task
+
+Fix qlty blocking issues on PR #740. User instruction: do not commit.
+
+## Investigation
+
+- (pending) Locate qlty config in repo.
+- (pending) Fetch qlty findings / CI status on PR #740.
+
+## Investigation results
+
+- qlty is used in CI only as a coverage uploader (`qlty-action/coverage`).
+  Code-quality analysis runs on the qlty cloud project.
+- Cloud `qlty check` status on PR head `0c7ec147` = SUCCESS (no blocking).
+- qlty dashboard requires auth; findings not fetchable via WebFetch.
+- Installed qlty CLI 0.637.0 and `qlty init` (auto-config, uncommitted).
+- Local `qlty check --upstream origin/master` medium+ findings are all
+  noise/out-of-scope:
+  - 136 `bandit:B101` (assert used) — all in test files (false positive).
+  - 25 `mypy:call-arg` + 2 `mypy:attr-defined` — qlty's mypy ignores the
+    project mypy/attrs config; project mypy + pyre are clean.
+  - 1 `python:S1186` empty method in unrelated `labels_validator.py`.
+- Conclusion: local auto-config diverges from cloud config; "fixing" these
+  would mean sprinkling suppressions / mangling attrs calls — wrong.
+  Need the actual cloud blocking findings to proceed.
+
+## Actual findings (from user)
+
+- 5 `qlty:similar-code` issues (currently ignored, hence check = SUCCESS),
+  all in `ebl/tests/fragmentarium/test_fragment_updater.py`. Two shown:
+  `test_update_introduction` and `test_update_notes` (mass 158).
+
+## Fix
+
+- Gitignored `.qlty/` (all local qlty files).
+- Extracted repeated changelog scaffold into
+  `ebl/tests/fragmentarium/fragment_updater_test_helpers.py`
+  (`entry`, `expect_changelog`).
+- Merged near-identical test families via `pytest.parametrize` (all cases
+  preserved): genres/date/dates_in_text -> `test_update_metadata_field`;
+  introduction/notes -> `test_update_edition_metadata_field`.
+- Used `expect_changelog` across the remaining updater tests.
+- `test_fragment_updater.py`: 417 -> 340 lines.
+
+## Gates
+
+- `task format`, flake8, mypy (changed files; 1 pre-existing
+  `parse_atf_lark` attr-defined, also in original), pyre: green.
+- Helper module coverage: 100%.
+- `task test`: 3933 passed, 2 skipped, 1 xfailed (unchanged count).
+
+## Notes
+
+- Local qlty config diverges from cloud and does not reproduce the 5
+  similar-code findings; final verification is the cloud `qlty check`
+  on push. The refactor removes the duplicated structures qlty flagged.
+- Not committed (per user instruction).
