+++ TASK-740-todo.md
@@ -1,0 +1,11 @@
+# TASK-740 TODO — Fix qlty blocking issues
+
+Scope: fix qlty blocking issues on PR #740. Do NOT commit.
+
+- [ ] Locate qlty config and determine what "blocking" means here.
+- [ ] Identify the specific qlty blocking findings (CI check and/or PR
+  annotations on PR #740).
+- [ ] Fix each blocking issue.
+- [ ] Re-verify with qlty locally (if available) and the standard gates.
+- [ ] Do not commit (per user instruction).
+- [ ] Remind to remove TASK docs before merge.
