+++ ebl/tests/fragmentarium/test_fragment_updater.py
@@ -1,30 +1,23 @@
-from typing import cast
-
 from ebl.common.domain.scopes import Scope
-from ebl.dictionary.domain.word import WordId
 from freezegun import freeze_time
 import pytest
 
-from ebl.errors import DataError, NotFoundError
-from ebl.fragmentarium.application.fragment_schema import FragmentSchema
-from ebl.fragmentarium.application.fragment_updater import FragmentUpdater
-from ebl.fragmentarium.domain.fragment import Fragment, Genre, NotLowestJoinError
+from ebl.errors import NotFoundError
+from ebl.fragmentarium.domain.fragment import Genre, NotLowestJoinError
 from ebl.fragmentarium.domain.joins import Join, Joins
 from ebl.transliteration.domain.museum_number import MuseumNumber
 from ebl.fragmentarium.domain.transliteration_update import TransliterationUpdate
-from ebl.lemmatization.domain.lemmatization import Lemmatization, LemmatizationToken
-from ebl.tests.factories.bibliography import ReferenceFactory
 from ebl.tests.factories.fragment import (
     FragmentFactory,
     TransliteratedFragmentFactory,
     DateFactory,
 )
+from ebl.tests.fragmentarium.fragment_updater_test_helpers import (
+    FROZEN_TIME,
+    expect_changelog,
+)
 from ebl.transliteration.domain.atf import Atf
 from ebl.transliteration.domain.atf_parsers.lark_parser import parse_atf_lark
-
-
-SCHEMA = FragmentSchema()
-FROZEN_TIME = "2018-09-07 15:41:24.032"
 
 
 @freeze_time(FROZEN_TIME)
@@ -61,12 +54,9 @@ def test_update_edition(
         .query_by_museum_number(number)
         .thenReturn(transliterated_fragment)
     )
-    when(changelog).create(
-        "fragments",
-        user.profile,
-        {"_id": str(number), **cast(dict, SCHEMA.dump(transliterated_fragment))},
-        {"_id": str(number), **cast(dict, SCHEMA.dump(transliterated_fragment))},
-    ).thenReturn()
+    expect_changelog(
+        when, changelog, user, number, transliterated_fragment, transliterated_fragment
+    )
     (
         when(fragment_repository)
         .update_field("transliteration", transliterated_fragment)
@@ -123,26 +113,35 @@ def test_update_update_transliteration_not_lowest_join(
         )
 
 
-def test_update_genres(
-    fragment_updater, user, fragment_repository, parallel_line_injector, changelog, when
+@pytest.mark.parametrize(
+    "field,value",
+    [
+        ("genres", (Genre(["ARCHIVAL", "Administrative"], False),)),
+        ("date", DateFactory.build()),
+        ("dates_in_text", [DateFactory.build()]),
+    ],
+)
+def test_update_metadata_field(
+    field,
+    value,
+    fragment_updater,
+    user,
+    fragment_repository,
+    parallel_line_injector,
+    changelog,
+    when,
 ):
     fragment = FragmentFactory.build()
     number = fragment.number
-    genres = (Genre(["ARCHIVAL", "Administrative"], False),)
-    updated_fragment = fragment.set_genres(genres)
+    updated_fragment = getattr(fragment, f"set_{field}")(value)
     injected_fragment = updated_fragment.set_text(
         parallel_line_injector.inject_transliteration(updated_fragment.text)
     )
     when(fragment_repository).query_by_museum_number(number).thenReturn(fragment)
-    when(changelog).create(
-        "fragments",
-        user.profile,
-        {"_id": str(number), **cast(dict, SCHEMA.dump(fragment))},
-        {"_id": str(number), **cast(dict, SCHEMA.dump(updated_fragment))},
-    ).thenReturn()
-    when(fragment_repository).update_field("genres", updated_fragment).thenReturn()
+    expect_changelog(when, changelog, user, number, fragment, updated_fragment)
+    when(fragment_repository).update_field(field, updated_fragment).thenReturn()
 
-    result = fragment_updater.update_genres(number, genres, user)
+    result = getattr(fragment_updater, f"update_{field}")(number, value, user)
     assert result == (injected_fragment, False)
 
 
@@ -162,256 +161,4 @@ def test_update_scopes(
     ).thenReturn()
     result = fragment_updater.update_scopes(number, scopes)
 
-    assert result == (injected_fragment, False)
-
-
-def test_update_date(
-    fragment_updater, user, fragment_repository, parallel_line_injector, changelog, when
-):
-    fragment = FragmentFactory.build()
-    number = fragment.number
-    date = DateFactory.build()
-    updated_fragment = fragment.set_date(date)
-    injected_fragment = updated_fragment.set_text(
-        parallel_line_injector.inject_transliteration(updated_fragment.text)
-    )
-    when(fragment_repository).query_by_museum_number(number).thenReturn(fragment)
-    when(changelog).create(
-        "fragments",
-        user.profile,
-        {"_id": str(number), **cast(dict, SCHEMA.dump(fragment))},
-        {"_id": str(number), **cast(dict, SCHEMA.dump(updated_fragment))},
-    ).thenReturn()
-    when(fragment_repository).update_field("date", updated_fragment).thenReturn()
-
-    result = fragment_updater.update_date(number, date, user)
-    assert result == (injected_fragment, False)
-
-
-def test_update_dates_in_text(
-    fragment_updater, user, fragment_repository, parallel_line_injector, changelog, when
-):
-    fragment = FragmentFactory.build()
-    number = fragment.number
-    dates_in_text = [DateFactory.build()]
-    updated_fragment = fragment.set_dates_in_text(dates_in_text)
-    injected_fragment = updated_fragment.set_text(
-        parallel_line_injector.inject_transliteration(updated_fragment.text)
-    )
-    when(fragment_repository).query_by_museum_number(number).thenReturn(fragment)
-    when(changelog).create(
-        "fragments",
-        user.profile,
-        {"_id": str(number), **cast(dict, SCHEMA.dump(fragment))},
-        {"_id": str(number), **cast(dict, SCHEMA.dump(updated_fragment))},
-    ).thenReturn()
-    when(fragment_repository).update_field(
-        "dates_in_text", updated_fragment
-    ).thenReturn()
-
-    result = fragment_updater.update_dates_in_text(number, dates_in_text, user)
-    assert result == (injected_fragment, False)
-
-
-@freeze_time(FROZEN_TIME)
-def test_update_lemmatization(
-    fragment_updater, user, fragment_repository, parallel_line_injector, changelog, when
-):
-    transliterated_fragment = TransliteratedFragmentFactory.build()
-    number = transliterated_fragment.number
-    tokens = [list(line) for line in transliterated_fragment.text.lemmatization.tokens]
-    tokens[1][3] = LemmatizationToken(tokens[1][3].value, (WordId("aklu I"),))
-    lemmatization = Lemmatization(tokens)
-    lemmatized_fragment = transliterated_fragment.update_lemmatization(lemmatization)
-    (
-        when(fragment_repository)
-        .query_by_museum_number(number)
-        .thenReturn(transliterated_fragment)
-    )
-    injected_fragment = lemmatized_fragment.set_text(
-        parallel_line_injector.inject_transliteration(lemmatized_fragment.text)
-    )
-    when(changelog).create(
-        "fragments",
-        user.profile,
-        {"_id": str(number), **cast(dict, SCHEMA.dump(transliterated_fragment))},
-        {"_id": str(number), **cast(dict, SCHEMA.dump(lemmatized_fragment))},
-    ).thenReturn()
-    when(fragment_repository).update_field(
-        "lemmatization", lemmatized_fragment
-    ).thenReturn()
-
-    result = fragment_updater.update_lemmatization(number, lemmatization, user)
-    assert result == (injected_fragment, False)
-
-
-def test_update_update_lemmatization_not_found(
-    fragment_updater, user, fragment_repository, when
-):
-    number = "K.1"
-    (when(fragment_repository).query_by_museum_number(number).thenRaise(NotFoundError))
-
-    with pytest.raises(NotFoundError):
-        fragment_updater.update_lemmatization(
-            number, Lemmatization(((LemmatizationToken("1.", ()),),)), user
-        )
-
-
-def test_update_references(
-    fragment_updater,
-    bibliography,
-    user,
-    fragment_repository,
-    parallel_line_injector,
-    changelog,
-    when,
-):
-    fragment = FragmentFactory.build()
-    number = fragment.number
-    reference = ReferenceFactory.build()
-    references = (reference,)
-    updated_fragment = fragment.set_references(references)
-    injected_fragment = updated_fragment.set_text(
-        parallel_line_injector.inject_transliteration(updated_fragment.text)
-    )
-    when(bibliography).find(reference.id).thenReturn(reference)
-    when(fragment_repository).query_by_museum_number(number).thenReturn(
-        fragment
-    ).thenReturn(updated_fragment)
-    when(fragment_repository).update_field("references", updated_fragment).thenReturn()
-    when(changelog).create(
-        "fragments",
-        user.profile,
-        {"_id": str(number), **cast(dict, SCHEMA.dump(fragment))},
-        {"_id": str(number), **cast(dict, SCHEMA.dump(updated_fragment))},
-    ).thenReturn()
-
-    result = fragment_updater.update_references(number, references, user)
-    assert result == (injected_fragment, False)
-
-
-def test_update_references_invalid(
-    fragment_updater, bibliography, user, fragment_repository, when
-):
-    fragment = FragmentFactory.build()
-    number = fragment.number
-    reference = ReferenceFactory.build()
-    when(bibliography).find(reference.id).thenRaise(NotFoundError)
-    (when(fragment_repository).query_by_museum_number(number).thenReturn(fragment))
-    references = (reference,)
-
-    with pytest.raises(DataError):
-        fragment_updater.update_references(number, references, user)
-
-
-def test_update_introduction(
-    fragment_updater: FragmentUpdater, user, fragment_repository, changelog, when
-):
-    fragment: Fragment = FragmentFactory.build()
-    number = fragment.number
-    introduction = "Test introduction"
-    updated_fragment = fragment.set_introduction(introduction)
-    when(fragment_repository).query_by_museum_number(number).thenReturn(fragment)
-    when(changelog).create(
-        "fragments",
-        user.profile,
-        {"_id": str(number), **cast(dict, SCHEMA.dump(fragment))},
-        {"_id": str(number), **cast(dict, SCHEMA.dump(updated_fragment))},
-    ).thenReturn()
-    when(fragment_repository).update_field(
-        "introduction", updated_fragment
-    ).thenReturn()
-
-    result = fragment_updater.update_edition(number, user, introduction=introduction)
-    assert result == (updated_fragment, False)
-
-
-def test_update_notes(
-    fragment_updater: FragmentUpdater, user, fragment_repository, changelog, when
-):
-    fragment: Fragment = FragmentFactory.build()
-    number = fragment.number
-    notes = "Test notes"
-    updated_fragment = fragment.set_notes(notes)
-    when(fragment_repository).query_by_museum_number(number).thenReturn(fragment)
-    when(changelog).create(
-        "fragments",
-        user.profile,
-        {"_id": str(number), **cast(dict, SCHEMA.dump(fragment))},
-        {"_id": str(number), **cast(dict, SCHEMA.dump(updated_fragment))},
-    ).thenReturn()
-    when(fragment_repository).update_field("notes", updated_fragment).thenReturn()
-
-    result = fragment_updater.update_edition(number, user, notes=notes)
-    assert result == (updated_fragment, False)
-
-
-@freeze_time(FROZEN_TIME)
-def test_update_lemma_annotation(
-    fragment_updater, user, fragment_repository, parallel_line_injector, changelog, when
-):
-    transliterated_fragment = TransliteratedFragmentFactory.build()
-    number = transliterated_fragment.number
-
-    annotation = {1: {3: ["aklu I"]}}
-    lemmatized_fragment = transliterated_fragment.update_lemma_annotation(annotation)
-
-    (
-        when(fragment_repository)
-        .query_by_museum_number(number)
-        .thenReturn(transliterated_fragment)
-    )
-    injected_fragment = lemmatized_fragment.set_text(
-        parallel_line_injector.inject_transliteration(lemmatized_fragment.text)
-    )
-    when(changelog).create(
-        "fragments",
-        user.profile,
-        {"_id": str(number), **cast(dict, SCHEMA.dump(transliterated_fragment))},
-        {"_id": str(number), **cast(dict, SCHEMA.dump(lemmatized_fragment))},
-    ).thenReturn()
-    when(fragment_repository).update_field(
-        "lemmatization", lemmatized_fragment
-    ).thenReturn()
-
-    result = fragment_updater.update_lemma_annotation(number, annotation, user)
-    assert result == (injected_fragment, False)
-
-
-@freeze_time(FROZEN_TIME)
-def test_update_named_entities(
-    fragment_updater,
-    named_entity_spans,
-    user,
-    fragment_repository,
-    parallel_line_injector,
-    changelog,
-    when,
-):
-    transliterated_fragment: Fragment = TransliteratedFragmentFactory.build()
-    number = transliterated_fragment.number
-
-    annotated_fragment = transliterated_fragment.set_named_entities(named_entity_spans)
-
-    (
-        when(fragment_repository)
-        .query_by_museum_number(number)
-        .thenReturn(transliterated_fragment)
-    )
-    injected_fragment = annotated_fragment.set_text(
-        parallel_line_injector.inject_transliteration(annotated_fragment.text)
-    )
-    when(changelog).create(
-        "fragments",
-        user.profile,
-        {"_id": str(number), **cast(dict, SCHEMA.dump(transliterated_fragment))},
-        {"_id": str(number), **cast(dict, SCHEMA.dump(annotated_fragment))},
-    ).thenReturn()
-    when(fragment_repository).update_field(
-        "named_entities", annotated_fragment
-    ).thenReturn()
-
-    result = fragment_updater.update_named_entities(
-        number, named_entity_spans, [], user
-    )
     assert result == (injected_fragment, False)
